Added unstable Graph Structure.

+++ Algorithms.cpp
@@ -26,7 +26,7 @@ namespace algo {
 	}
 	int& getMax(int* arr, size_t size) {
 		int max = arr[0];
-		for (int i = 1; i < size; i++)
+		for (size_t i = 1; i < size; i++)
 			if (arr[i] > max)
 				max = arr[i];
 		return max;
